+++ lib/oauth.dart
@@ -2,4 +2,4 @@ library oauth;
 
 export 'src/BraspagOAuth.dart';
 export 'src/ErrorResponseOAuth.dart';
-export 'src/OAuthEnviroment.dart';
+export 'src/OAuthEnvironment.dart';