+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 - Improved callback error.
 - Import package name changed
 - Improved readme.
+
+## [1.1.1] - 02/10/2020
+
+- Fixed Typo.
+- Adjustment in tests
+- Adjustment in error return.
+